+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 - [CSE] Added validation of received response, e.g. in transfer rerequests. 
 - [CSE] Added a minimal console with only a few basic commands. This can be used when the CSE is running in headless mode or where access to the console is not possible.
 - [CSE] Added support for new *spi* and *ici* attributes in &lt;CSEBsae> and &lt;remoteCSE> resources. Both attributes are updated automatically in registree CSEs throughout a oneM2M deployment tree. This is an ESTIMED project contribution.
+- [CSE] Added support for decorating event handlers.
 - [WEB] Added support for external root path for the web UI. This allows to run the web UI behind a reverse proxy under a specific path or a Kubernetes ingress path. Thanks to Mudassar Khan (xFlow) for the idea. This is an ESTIMED project contribution.
 - [TUI] Added support for displaying the originator of a resource in the text UI. The originator is now shown in the resource subtitle in the resource view. Clicking on it copies the originator to the clipboard.